+++ api/src/main/java/org/allaymc/api/eventbus/event/player/PlayerUseItemInAirEvent.java
@@ -1,0 +1,26 @@
+package org.allaymc.api.eventbus.event.player;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.allaymc.api.entity.interfaces.EntityPlayer;
+import org.allaymc.api.eventbus.event.CancellableEvent;
+
+/**
+ * Called when a player used an item in air.
+ *
+ * @author daoge_cmd
+ */
+@Getter
+@Setter
+public class PlayerUseItemInAirEvent extends PlayerEvent implements CancellableEvent {
+
+    /**
+     * The time the player has been using the item in air in game ticks.
+     */
+    protected long usingTime;
+
+    public PlayerUseItemInAirEvent(EntityPlayer player, long usingTime) {
+        super(player);
+        this.usingTime = usingTime;
+    }
+}