+++ server/src/main/java/org/allaymc/server/item/component/ItemArmorBaseComponentImpl.java
@@ -34,7 +34,7 @@ public class ItemArmorBaseComponentImpl extends ItemBaseComponentImpl implements
     }
 
     @Override
-    public void clickItemInAir(EntityPlayer player) {
+    public void rightClickItemInAir(EntityPlayer player) {
         equipArmor(player, thisItemStack);
     }