+++ server/src/main/java/org/allaymc/server/item/component/ItemAxeBaseComponentImpl.java
@@ -25,8 +25,8 @@ public class ItemAxeBaseComponentImpl extends ItemBaseComponentImpl {
     }
 
     @Override
-    public void rightClickItemOn(Dimension dimension, Vector3ic placeBlockPos, PlayerInteractInfo interactInfo) {
-        super.rightClickItemOn(dimension, placeBlockPos, interactInfo);
+    public void rightClickItemOnBlock(Dimension dimension, Vector3ic placeBlockPos, PlayerInteractInfo interactInfo) {
+        super.rightClickItemOnBlock(dimension, placeBlockPos, interactInfo);
 
         var clickedBlockPos = interactInfo.clickedBlockPos();
         var clickedBlockState = dimension.getBlockState(clickedBlockPos);